+++ eras/mary/impl/test/Test/Cardano/Ledger/Mary/BinarySpec.hs
@@ -9,6 +9,7 @@ import Cardano.Ledger.Mary
 import Data.Default (def)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Binary (specUpgrade)
+import Test.Cardano.Ledger.Core.Binary as Binary (decoderEquivalenceCoreEraTypesSpec)
 import Test.Cardano.Ledger.Core.Binary.RoundTrip (RuleListEra (..))
 import Test.Cardano.Ledger.Mary.Arbitrary ()
 import Test.Cardano.Ledger.Mary.TreeDiff ()
@@ -19,6 +20,8 @@ spec = do
   specUpgrade @MaryEra def
   describe "RoundTrip" $ do
     roundTripShelleyCommonSpec @MaryEra
+  describe "DecCBOR instances equivalence" $ do
+    Binary.decoderEquivalenceCoreEraTypesSpec @MaryEra
 
 instance RuleListEra MaryEra where
   type